requirements.txt
asgiref
attrs
dataclasses
Django
djangorestframework==3.15.2
dnspython
drf-spectacular==0.27.2
drf-yasg==1.21.8
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.1
pymongo
python-decouple==3.8
pytz
PyYAML
referencing==0.35.1
rpds-py==0.20.0
sqlparse
tzdata
uritemplate
